+++ web/src/hotspots.js
@@ -7,6 +7,7 @@ import render_hotspot_overlay from "../templates/hotspot_overlay.hbs";
 
 import * as blueslip from "./blueslip";
 import * as channel from "./channel";
+import * as overlays from "./overlays";
 import {page_params} from "./page_params";
 import * as popovers from "./popovers";
 
@@ -271,4 +272,55 @@ export function load_new(new_hotspots) {
 
 export function initialize() {
     load_new(page_params.hotspots);
+
+    // open
+    $("body").on("click", ".hotspot-icon", function (e) {
+        // hide icon
+        close_hotspot_icon(this);
+
+        // show popover
+        const [, hotspot_name] = /^hotspot_(.*)_icon$/.exec(
+            $(e.target).closest(".hotspot-icon").attr("id"),
+        );
+        const overlay_name = "hotspot_" + hotspot_name + "_overlay";
+
+        overlays.open_overlay({
+            name: overlay_name,
+            $overlay: $(`#${CSS.escape(overlay_name)}`),
+            on_close: function () {
+                // close popover
+                $(this).css({display: "block"});
+                $(this).animate(
+                    {opacity: 1},
+                    {
+                        duration: 300,
+                    },
+                );
+            }.bind(this),
+        });
+
+        e.preventDefault();
+        e.stopPropagation();
+    });
+
+    // confirm
+    $("body").on("click", ".hotspot.overlay .hotspot-confirm", function (e) {
+        e.preventDefault();
+        e.stopPropagation();
+
+        const overlay_name = $(this).closest(".hotspot.overlay").attr("id");
+
+        const [, hotspot_name] = /^hotspot_(.*)_overlay$/.exec(overlay_name);
+
+        // Comment below to disable marking hotspots as read in production
+        post_hotspot_as_read(hotspot_name);
+
+        overlays.close_overlay(overlay_name);
+        $(`#hotspot_${CSS.escape(hotspot_name)}_icon`).remove();
+    });
+
+    // stop propagation
+    $("body").on("click", ".hotspot.overlay .hotspot-popover", (e) => {
+        e.stopPropagation();
+    });
 }
